Made specific classes for webToJson and JsonToDB and cleaned up the code with additional comments
