+++ src/main/java/FootballLeague/FootballLeagueFrontend/GameMenu.java
@@ -265,8 +265,9 @@ public class GameMenu extends Application {
             String formattedDate = myDateObj.format(myFormatObj);
             String jsonFileName = dbSize.trim() + formattedDate;
 
-            Task task = returnTaskJsonCreation(dbSize, jsonFileName);
+            Task task = WebToJson.returnTaskJsonCreation(dbSize, jsonFileName);
 
+            //Used to bind the progress bar to the progress of the task
             progressBox.progressBar.progressProperty().bind(task.progressProperty());
 
             //Code to be run once the task has been completed successfully
@@ -417,258 +418,5 @@ public class GameMenu extends Application {
 
     public static void main(String[] args) {
         launch(args);
-    }
-
-    //~~~~~~~~ Methods used for web scraping ~~~~~~~~~~~//
-
-    public static ArrayList<String> getPlayerLinks(String teamUrl){
-        WebClient client = new WebClient();
-        client.getOptions().setCssEnabled(false);
-        client.getOptions().setJavaScriptEnabled(false);
-        List<HtmlElement> playerRows = new ArrayList<>();
-        try{
-            HtmlPage page = client.getPage(teamUrl);
-            playerRows = page.getByXPath("//*[@id=\"yw1\"]/table/tbody/tr");
-        } catch (Exception e){
-            e.printStackTrace();
-        }
-        ArrayList<String> playerLinks = new ArrayList<>();
-        if(playerRows.isEmpty()){
-            System.out.println("No Results");
-        } else {
-            for(HtmlElement row : playerRows){
-                String profileHtml = ((HtmlAnchor)row.getFirstByXPath("./td/table/tbody/tr/td/div/span/a")).getHrefAttribute();
-                playerLinks.add(profileHtml);
-            }
-        }
-        return playerLinks;
-    }
-
-    public static ArrayList<String> getTeamLinks(String leagueUrl){
-        System.out.println("Starting to get team links");
-        WebClient client = new WebClient();
-        client.getOptions().setCssEnabled(false);
-        client.getOptions().setJavaScriptEnabled(false);
-        ArrayList<String> teamLinks = new ArrayList<>();
-
-        try{
-            HtmlPage page = client.getPage(leagueUrl);
-            for(int i=1; i<30; i++){
-                if((page.getFirstByXPath("//*[@id=\"yw1\"]/table/tbody/tr[" + i + "]/td[2]/a")) == null){
-                    i = 50;
-                } else {
-                    String link = ((HtmlAnchor) page.getFirstByXPath("//*[@id=\"yw1\"]/table/tbody/tr[" + i + "]/td[2]/a")).getHrefAttribute();
-                    teamLinks.add(link);
-                }
-
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        System.out.println("Finished getting the team links");
-        return teamLinks;
-    }
-
-    public static TeamJsonScraper getTeam(String teamUrl){
-
-        try{
-            WebClient client = new WebClient();
-            client.getOptions().setCssEnabled(false);
-            client.getOptions().setJavaScriptEnabled(false);
-            HtmlPage page = client.getPage(teamUrl);
-
-            String name = ((HtmlElement) page.getFirstByXPath("//*[@id=\"verein_head\"]/div/div[1]/div[1]/div/div[1]/h1/span")).asText();
-            String stadiumName = ((HtmlElement) page.getFirstByXPath("//*[@id=\"verein_head\"]/div/div[1]/div[2]/div/div[2]/p[2]/span[2]/a")).asText();
-            String stadiumCapacityString = ((HtmlElement) page.getFirstByXPath("//*[@id=\"verein_head\"]/div/div[1]/div[2]/div/div[2]/p[2]/span[2]/span")).asText().replace(".", "");
-            String league = ((HtmlImage) page.getFirstByXPath("//*[@id=\"verein_head\"]/div/div[1]/div[4]/div[1]/a/img")).getAltAttribute();
-
-            int stadiumCapacity = Integer.parseInt(stadiumCapacityString.split(" ")[0]);
-
-            return new TeamJsonScraper(name, stadiumName, stadiumCapacity, league);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return new TeamJsonScraper();
-    }
-
-
-    public static PlayerJsonScraper getPlayer(String playerUrl) {
-        PlayerJsonScraper player = new PlayerJsonScraper();
-
-        try{
-            WebClient client = new WebClient();
-            client.getOptions().setCssEnabled(false);
-            client.getOptions().setJavaScriptEnabled(false);
-            HtmlPage page = client.getPage(playerUrl);
-
-            String dateOfBirth = null;
-            int age = 0;
-            String position = null;
-            String nationality = null;
-            String team = null;
-
-            //Iterate over the table headers
-            for(int i=1; i < 20; i++){
-                HtmlElement tableData = null;
-                HtmlElement tableHeader = null;
-                if((page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[2]/div[2]/div[2]/div[2]/table/tbody/tr[" + i + "]/td")) == null){
-                    tableData = (page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[1]/div[2]/div[2]/div[2]/table/tbody/tr[" + i + "]/td"));
-                    tableHeader=(page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[1]/div[2]/div[2]/div[2]/table/tbody/tr[" + i + "]/th"));
-                } else {
-                    tableData = (page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[2]/div[2]/div[2]/div[2]/table/tbody/tr[" + i + "]/td"));
-                    tableHeader=(page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[2]/div[2]/div[2]/div[2]/table/tbody/tr[" + i + "]/th"));
-                }
-                switch(tableHeader.asText()){
-                    case "Date of birth:":
-                        dateOfBirth = tableData.asText();
-                        break;
-                    case "Age:":
-                        age = Integer.parseInt(tableData.asText());
-                        break;
-                    case "Citizenship:":
-                        nationality = ((HtmlImage) tableData.getFirstByXPath("//td/img")).getAltAttribute();
-                        break;
-                    case "Position:":
-                        position = tableData.asText();
-                        break;
-                    case "Current club:":
-                        team = ((HtmlElement) tableData.getFirstByXPath("//td/a[2]")).asText();;
-                        i = 21;
-                        break;
-                }
-            }
-
-            String name;
-            if((page.getFirstByXPath("//*[@id=\"main\"]/div[8]/div/div/div[1]/div/div/h1")) == null){
-                name = ((HtmlElement) page.getFirstByXPath("//*[@id=\"main\"]/div[8]/div/div/div[2]/div/div/h1")).asText();
-            } else {
-                name = ((HtmlElement) page.getFirstByXPath("//*[@id=\"main\"]/div[8]/div/div/div[1]/div/div/h1")).asText();
-            }
-
-            long value = 0L;
-            if(page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[2]/div[2]/div[1]/div/div[4]/div[1]/div[1]/div/div[1]/div[2]/a[1]") != null){
-                value = formatValue(((HtmlElement) page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[2]/div[2]/div[1]/div/div[4]/div[1]/div[1]/div/div[1]/div[2]/a[1]")).asText().replace(".",""));
-            }
-            if(page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[1]/div[2]/div[1]/div/div[4]/div[1]/div[1]/div/div[1]/div[2]/a[1]") != null){
-                value = formatValue(((HtmlElement) page.getFirstByXPath("//*[@id=\"main\"]/div[11]/div[1]/div[1]/div[2]/div[1]/div/div[4]/div[1]/div[1]/div/div[1]/div[2]/a[1]")).asText().replace(".",""));
-            }
-
-            String yw = null;
-            assert position != null;
-            if(page.getFirstByXPath("//*[@id=\"yw2\"]/table/tfoot/tr/td[5]") != null){
-                yw = "yw2";
-            } else if(page.getFirstByXPath("//*[@id=\"yw1\"]/table/tfoot/tr/td[5]") != null){
-                yw = "yw1";
-            }
-
-            int gamesPlayed = stringToInt(((HtmlElement) page.getFirstByXPath("//*[@id=\"" + yw + "\"]/table/tfoot/tr/td[3]")).asText());
-
-            if(position.equals("Goalkeeper") && yw != null){
-                int cleanSheets = stringToInt(((HtmlElement) page.getFirstByXPath("//*[@id=\"" + yw + "\"]/table/tfoot/tr/td[5]")).asText());
-                int goalsConceded = stringToInt(((HtmlElement) page.getFirstByXPath("//*[@id=\"" + yw + "\"]/table/tfoot/tr/td[4]")).asText());
-                player = new PlayerJsonScraper(name, dateOfBirth, age, position, nationality, team, gamesPlayed, cleanSheets, goalsConceded, value, playerUrl);
-            } else if(yw != null) {
-                int goalsScored = stringToInt(((HtmlElement) page.getFirstByXPath("//*[@id=\"" + yw + "\"]/table/tfoot/tr/td[4]")).asText());
-                int assists = stringToInt(((HtmlElement) page.getFirstByXPath("//*[@id=\"" + yw + "\"]/table/tfoot/tr/td[5]")).asText());
-                player = new PlayerJsonScraper("", name, dateOfBirth, age, position, nationality, team, gamesPlayed, goalsScored, assists, value, playerUrl);
-            } else {
-                player = new PlayerJsonScraper(name, dateOfBirth, age, position, nationality, team, gamesPlayed, 0, 0, 0, 0, value, playerUrl);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        System.out.println("Completed - " + player.name);
-        return player;
-    }
-
-    public static int stringToInt(String str){
-        if(str.equals("-")){
-            return 0;
-        } else {
-            return Integer.parseInt(str);
-        }
-    }
-
-    public static long formatValue(String str){
-        str = str.substring(1);
-        if(str.contains("m")){
-            str = str.replace("m","");
-            return Long.parseLong(str) * 1000000 / 100;
-        }
-        if(str.contains("k")){
-            str = str.replace("k","");
-            return Long.parseLong(str) * 1000;
-        }
-        return Long.parseLong(str);
-    }
-
-    public Task returnTaskJsonCreation(String dbSize, String jsonFileName){
-        return new Task() {
-            @Override
-            protected Object call() throws Exception {
-                ArrayList<String> teamLinks = new ArrayList<>();
-                ArrayList<PlayerJsonScraper> playerList = new ArrayList<>();
-                ArrayList<TeamJsonScraper> teamList = new ArrayList<>();
-                System.out.println("Start Json Creation");
-                String baseUrl = "https://www.transfermarkt.co.uk";
-                String premierLeagueUrl = "https://www.transfermarkt.co.uk/premier-league/startseite/wettbewerb/GB1";
-                String championshipUrl = "https://www.transfermarkt.co.uk/championship/startseite/wettbewerb/GB2/saison_id/2019";
-                String bundesligaUrl = "https://www.transfermarkt.co.uk/1-bundesliga/startseite/wettbewerb/L1";
-                String laLigaUrl = "https://www.transfermarkt.co.uk/primera-division/startseite/wettbewerb/ES1";
-                String serieAUrl = "https://www.transfermarkt.co.uk/serie-a/startseite/wettbewerb/IT1";
-                String ligue1Url = "https://www.transfermarkt.co.uk/ligue-1/startseite/wettbewerb/FR1";
-
-                if(dbSize.equals("Small ")){
-                    //Takes ~17 minutes 48 seconds to complete
-                    teamLinks = getTeamLinks(premierLeagueUrl);
-                } else if(dbSize.equals("Medium ")){
-                    //Takes ~37 minutes 5 seconds to complete
-                    teamLinks = getTeamLinks(championshipUrl);
-                    teamLinks.addAll(getTeamLinks(premierLeagueUrl));
-                } else if(dbSize.equals("Large ")){
-                    //Takes ~1 hour 20 minutes 10 seconds to complete
-                    teamLinks = getTeamLinks(premierLeagueUrl);
-                    teamLinks.addAll(getTeamLinks(championshipUrl));
-                    teamLinks.addAll(getTeamLinks(bundesligaUrl));
-                    teamLinks.addAll(getTeamLinks(laLigaUrl));
-                    teamLinks.addAll(getTeamLinks(serieAUrl));
-                    teamLinks.addAll(getTeamLinks(ligue1Url));
-                }
-                int updateProgress = 1;
-                for (String teamLink : teamLinks) {
-                    if(isCancelled()){
-                        System.out.println("Is cancelled");
-                        break;
-                    }
-                    teamList.add(getTeam(baseUrl + teamLink));
-                    for (String playerLink : getPlayerLinks(baseUrl + teamLink)) {
-                        //These players have insufficient data to use so will be skipped
-                        if( playerLink.equals("/ahmadou-dia/profil/spieler/611673") ||
-                                playerLink.equals("/callum-pearson/profil/spieler/547036") ||
-                                playerLink.equals("/eiji-kawashima/profil/spieler/77383")  ||
-                                playerLink.equals("/anthony-rouault/profil/spieler/690799")  ||
-                                playerLink.equals("/boris-essele/profil/spieler/694499")  ||
-                                playerLink.equals("/darell-tokpa/profil/spieler/655414")  ||
-                                playerLink.equals("/louis-pelletier/profil/spieler/616283")) {
-
-                        }
-                        else {
-                            playerList.add(getPlayer(baseUrl + playerLink));
-                            System.out.println(playerLink);
-                        }
-                    }
-                    //Updates the progress value of the task which is linked to the progress bar
-                    updateProgress++;
-                    updateProgress(updateProgress, teamLinks.size());
-                }
-                try {
-                    //Writes the players to the json file
-                    JsonScraperOperations.writePlayer(teamList, playerList, jsonFileName);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                return null;
-            }
-        };
     }
 }
